requirements.txt
Flask==3.0.3
Flask_Login==0.6.3
Pillow==10.4.0
protobuf==5.27.2
PyDrive2==1.19.0
PyMuPDF==1.24.7
pymongo==4.8.0
pytesseract==0.3.10
python-dotenv==1.0.1
scholarly==1.7.11
Unidecode==1.3.8
Werkzeug==3.0.3
gunicorn==22.0.0